moving logic into RTMv5
