+++ fitpay/src/main/java/com/fitpay/android/webview/impl/parser/RtmParserV5.java
@@ -1,10 +1,15 @@
 package com.fitpay.android.webview.impl.parser;
 
+import com.fitpay.android.utils.Constants;
 import com.fitpay.android.utils.RxBus;
 import com.fitpay.android.webview.enums.RtmType;
 import com.fitpay.android.webview.events.IdVerificationRequest;
 import com.fitpay.android.webview.events.RtmMessage;
+import com.fitpay.android.webview.events.RtmMessageResponse;
+import com.fitpay.android.webview.events.a2a.A2AVerificationFailed;
+import com.fitpay.android.webview.events.a2a.A2AVerificationRequest;
 import com.fitpay.android.webview.impl.WebViewCommunicatorImpl;
+import com.fitpay.android.webview.models.a2a.A2ASupportsAuth;
 
 /**
  * RtmMessage parser v5
@@ -20,6 +25,21 @@ public class RtmParserV5 extends RtmParserV4 {
         switch (msg.getType()) {
             case RtmType.ID_VERIFICATION_REQUEST:
                 RxBus.getInstance().post(new IdVerificationRequest(msg.getCallbackId()));
+                break;
+
+            case RtmType.SUPPORTS_ISSUER_APP_AUTH:
+                RxBus.getInstance().post(new RtmMessageResponse(msg.getCallbackId(), true, new A2ASupportsAuth(impl.isSupportAppAuth()), RtmType.SUPPORTS_ISSUER_APP_AUTH));
+                break;
+
+            case RtmType.APP_TO_APP_VERIFICATION:
+                if (impl.isSupportAppAuth()) {
+                    A2AVerificationRequest appToAppVerification = Constants.getGson().fromJson(msg.getData(), A2AVerificationRequest.class);
+                    appToAppVerification.setCallbackId(msg.getCallbackId());
+                    RxBus.getInstance().post(appToAppVerification);
+                } else {
+                    RxBus.getInstance().post(new RtmMessageResponse(msg.getCallbackId(),false,
+                            new A2AVerificationFailed("a2a auth is not supported"), RtmType.APP_TO_APP_VERIFICATION));
+                }
                 break;
 
             default:
